requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-auth==2.29.0
google-api-python-client==2.126.0
pyyaml==6.0.1
gunicorn==22.0.0
